rename General.vue to ActivityContentLayout.vue
