+++ frontend/src/components/activity/ActivityContentLayout.vue
